client/requirements.txt
textual
python-dotenv
requests
python-socketio
eventlet